Created odin-recipe and added index and 2 recipes
I had reached the point where I am creating my first project just using
HTML and I am enjoying the process. So far the index linkes to two
different recipes that have the name, ingredients, and steps for the
recipe. The first recipe for the pasta is finished and I am currently
working on implementing the second recipe which is the chocolate
chip cookie recipe. This has helped me familiar myself more with
basic HTML and I look forward to working on this and other projects
in the future.