Added salsa.html file to recipes
After the previous day I decided to finish the project and added
the finale recipe to the recipe directory. This has been an interesting
project as I see it probably has given me the ability to really drill in
some of the different parts of HTML that I have learned so far. I am looking
forward to more of this work and eventually more back-end stuff.